Lifted all queue lifecycle management to `HoodieExecutorBase`
